{{cookiecutter.repo_name}}/requirements/local.txt
-r base.txt

coverage==3.6
django-discover-runner==0.4
django-debug-toolbar==1.0.1

